apps/llmsr-worker/pyproject.toml
[project]
name = "llmsr-worker"
version = "0.1.0"
description = "llmsr worker gRPC server"
readme = "README.md"
authors = [{ name = "Xo1otl", email = "xolotl.mictl4n@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.75.1",
    "grpcio-reflection>=1.75.1",
    "pytest>=8.4.2",
]

[project.scripts]
llmsr-worker = "llmsr_worker:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "grpcio-tools>=1.75.1",
]
